+++ src/rics/utility/pandas/__init__.py
@@ -1,0 +1,5 @@
+"""Utility functions for :mod:`pandas`."""
+
+from rics.utility.pandas._time_fold import TimeFold
+
+__all__ = ["TimeFold"]